pyproject.toml
[tool.poetry]
name = "ao"
version = "0.1.0"
description = ""
authors = ["Col Perks <wild.fauve@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
ao = "ao.cli:cli"


[tool.poetry.dependencies]
python = "^3.9"
PyMonad = "^2.4.0"
click = "^8.1.3"
rdflib = "^6.2.0"


[tool.poetry.group.dev.dependencies]
pytest-env = "^0.8.1"
pdbpp = "^0.10.3"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
python_classes = [
    "*Test",
    "Test*",
    "Describe*",
]
python_functions = [
    "test_*",
    "it_*",
]
xfail_strict = true
log_cli = true
log_cli_level = 20
env = [
]
testpaths = [
    "tests",
    "integration",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
